build: silence an annoying warning on Windows
This alteration to Package.swift mirrors the changes that were done in
the SwiftDocOrg/Markup package.  It removes the pkg-config, brew
providers to silence some warnings when building on Windows.  These
tools are normally not distributed on Windows, and require that the user
manually pass along the graphviz installation path.

This is preparatory work, there is further fixes required to enable the
use of this package on Windows.